platformio/various: cleaned up build flags
- Added flags to all builds:
    *fno-common
    *Wformat=2
    *Wformat-truncation
    *Wno-format-nonliteral ; needed for strftime with dynamic template
- Added -Os to release build
- Refactored some functions to fixate formatting across platforms
- Removed EMBEDDED flag as only NATIVE needs special treatment

+++ src/spine/core/utils/time_repr.hpp
@@ -4,6 +4,7 @@
 
 #include <array>
 // #include <charconv>
+#include <cinttypes>
 #include <cstring>
 #include <iterator>
 #include <numeric>
@@ -55,7 +56,7 @@ inline std::string repr(const TimeType& t) {
                                             : "unknown";
 
     /// todo: this is a workaround, see issue with charconv: https://github.com/s-t-a-n/Spine/issues/21
-    auto size = std::snprintf(buffer.data(), buffer.size(), "%i%s", raw, unit);
+    auto size = std::snprintf(buffer.data(), buffer.size(), "%" PRId64 "%s", static_cast<int64_t>(raw), unit);
     if (size < 0 || static_cast<size_t>(size) >= buffer.size()) {
         return "Error";
     }
